# eth/tracers: refactor exporting js buffer